Codec.Compression.Zstd.Lazy

Description

Lazy compression and decompression support for zstd. Under the hood, these are implemented using the streaming APIs.

Documentation

compress

Arguments

:: Int
Compression level. Must be >= 1 and <= maxCLevel.

-> ByteString
Payload to compress. This will be consumed lazily.

-> ByteString

Compress a payload. The input will be consumed lazily, and the compressed result generated lazily.

Note: if any error occurs, compression will fail part-way through with a call to error.

decompress :: ByteString -> ByteString

Decompress a payload. The input will be consumed lazily, and the decompressed result generated lazily.

Note: if any error occurs, decompression will fail part-way through with a call to error.

maxCLevel :: Int

The maximum compression level supported by the library.