Codec.Compression.Zstd.Types

Description

Types supporting zstd compression and decompression.

Documentation

data Decompress

The result of a decompression operation.

Constructors

Skip
Either the compressed frame was empty, or it was compressed in streaming mode and so its size is not known.

Error String
An error occurred.

Decompress ByteString
The payload was successfully decompressed.

newtype Dict

Compression dictionary.

Constructors

Dict

Fields

fromDict :: ByteString

mkDict :: ByteString -> Dict

Smart constructor.